requirements.txt
lightkurve>=2.4.0
astropy>=5.3.0
batman-package>=2.4.9
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
astroquery>=0.4.6
pandas>=2.0.0
tqdm>=4.65.0
